+++ .prettierrc.js
@@ -18,6 +18,7 @@ export default {
     '',
     '^@oxide/(.*)$',
     '',
+    '^~/(.*)$',
     '^app/(.*)$',
     '',
     '^[./]',
