+++ app/table/cells/SizeCell.tsx
