+++ app/table/cells/TruncateCell.tsx
