+++ app/table/cells/TypeValueCell.tsx
