+++ app/table/cells/TypeValueListCell.tsx
