+++ tsconfig.json
@@ -12,14 +12,14 @@
     "noEmit": true,
     "outDir": "dist",
     "paths": {
+      "~/*": ["app/*"],
       "app/*": ["app/*"],
       "@oxide/gen/*": ["libs/api/__generated__/*"],
       "@oxide/api": ["libs/api/index.ts"],
       "@oxide/api-mocks": ["libs/api-mocks/index.ts"],
       "@oxide/ui": ["libs/ui/index.ts"],
       "@oxide/util": ["libs/util/index.ts"],
-      "@oxide/table": ["libs/table/index.ts"],
-      "@oxide/pagination": ["libs/pagination/index.ts"]
+      "@oxide/table": ["app/table/index.ts"]
     },
     "resolveJsonModule": true,
     "rootDir": ".",
